process: Move types/command.go to process/
